requirements.txt
streamlit
python-dotenv
google-generativeai
langchain-chroma
langchain-google-genai
langchain-community
langchain-core
langchain-text-splitters
chromadb
pysqlite3